requirements.txt
mcp
langchain
langchain-openai
langchain-community
chromadb
sqlalchemy
pandas
sympy
rich
python-dotenv
pydantic